Use wheezy packages for unsupported Debian version
As of now, we do not support Debian jessie or newer on our apt repos.
This work adjusts the class to catch the release to see if we support
it.  A known newer-than-supported list of releases is downgraded to
Debian wheezy packages.

When we ship jessie packages, we can remove jessie from the list to
restore directly passing $release to the apt::source define.